added event of opening and closing popup with image
