+++ blocks/popup/__image-description/popup__image-description.css
@@ -1,0 +1,10 @@
+.popup__image-description {
+  font-family: Inter, sans-serif;
+  font-style: normal;
+  font-weight: 400;
+  font-size: 12px;
+  line-height: 15px;
+  color: white;
+  margin-top: 10px;
+  margin-bottom: 0;
+}
