+++ blocks/popup/_opened/popup_opened.css
@@ -1,5 +1,5 @@
 .popup_opened {
   visibility: visible;
   opacity: 1;
-  transition: visibility 0.6s, opacity 0.5s linear;
+  transition: visibility 0.6s, opacity 0.6s linear;
 }
