+++ blocks/popup/popup.css
@@ -11,5 +11,5 @@
   display: flex;
   visibility: hidden;
   opacity: 0;
-  transition: visibility 0.6s, opacity 0.5s linear;
+  transition: visibility 0.6s, opacity 0.6s linear;
 }
