+++ index.html
@@ -29,54 +29,6 @@
         </section>
         <section class="elements">
           <p class="elements__no-items">Добавьте своё первое место!</p>
-         <!-- <figure class="element">
-            <img class="element__image" src="images/elbrus.jpg" alt="место">
-            <figcaption class="element__caption">
-              <h3 class="element__caption-text">Эльбрус</h3>
-              <button type="button" class="element__heart"></button>
-            </figcaption>
-            <button class="element__delete-btn" type="button"></button>
-          </figure>
-          <figure class="element">
-            <img class="element__image" src="images/kishi.jpg" alt="место">
-            <figcaption class="element__caption">
-              <h3 class="element__caption-text">Кижи</h3>
-              <button type="button" class="element__heart"></button>
-            </figcaption>
-            <button class="element__delete-btn" type="button"></button>
-          </figure>
-          <figure class="element">
-            <img class="element__image" src="images/baikal.jpg" alt="место">
-            <figcaption class="element__caption">
-              <h3 class="element__caption-text">Озеро Байкал</h3>
-              <button type="button" class="element__heart"></button>
-            </figcaption>
-            <button class="element__delete-btn" type="button"></button>
-          </figure>
-          <figure class="element">
-            <img class="element__image" src="images/seliger.jpeg" alt="место">
-            <figcaption class="element__caption">
-              <h3 class="element__caption-text">Озеро Селигер</h3>
-              <button type="button" class="element__heart"></button>
-            </figcaption>
-            <button class="element__delete-btn" type="button"></button>
-          </figure>
-          <figure class="element">
-            <img class="element__image" src="images/altai.jpg" alt="место">
-            <figcaption class="element__caption">
-              <h3 class="element__caption-text">Алтайские горы</h3>
-              <button type="button" class="element__heart"></button>
-            </figcaption>
-            <button class="element__delete-btn" type="button"></button>
-          </figure>
-          <figure class="element">
-            <img class="element__image" src="images/lensky.jpg" alt="место">
-            <figcaption class="element__caption">
-              <h3 class="element__caption-text">Ленские столбы</h3>
-              <button type="button" class="element__heart"></button>
-            </figcaption>
-            <button class="element__delete-btn" type="button"></button>
-          </figure>-->
         </section>
       </main>
       <footer class="footer">
@@ -99,10 +51,9 @@
         <section class="popup__container">
           <button class="popup__close-button" type="button"></button>
           <img class="popup__image" src="#" alt="Фото места">
-          <p class="image__description"></p>
+          <p class="popup__image-description"></p>
         </section>
       </div>
-
       <template id="element-template">
         <figure class="element">
           <img class="element__image" src="#" alt="место">
@@ -114,7 +65,8 @@
         </figure>
       </template>
     </div>
-    <script src="./scripts/index.js" type="module"></script>
+    <script src="./scripts/popup.js" type="module"></script>
     <script src="./scripts/template.js" type="module"></script>
+    <script src="./scripts/index.js" type="module"></script>
   </body>
 </html>
