+++ pages/index.css
@@ -39,6 +39,7 @@
 @import url(../blocks/popup/__container/popup__container.css);
 @import url(../blocks/popup/__close-button/popup__close-button.css);
 @import url(../blocks/popup/__image/popup__image.css);
+@import url(../blocks/popup/__image-description/popup__image-description.css);
 @import url(../blocks/popup/_with/popup_with_image.css);
 
 /*form*/
